parser/markdown: markdown parser refactoring
The markdown parser wasn't reentrant, so list in list, italic text in a list
or note, and other similar cases weren't supported.
Also, some objects like survey and others weren't supported either.

Use the gdoc parser thus to mirror the functionality and global logic. Adapt
it as we don't have css and the output from blackfriday conversion is a little
bit different..